+++ src/api/init.ts
@@ -2,12 +2,9 @@ import { version } from '../../package.json';
 import * as websocket from '../ws/websocket';
 import * as debug from './debug';
 import * as events from './events';
+import type { InitOptions } from '../types/api/init';
 
 let initialized = false;
-
-export interface InitOptions {
-    exportCustomMethods?: boolean;
-}
 
 export function init(options: InitOptions = {}): void {
     options = { exportCustomMethods: true ,...options };
@@ -26,10 +23,10 @@ export function init(options: InitOptions = {}): void {
     }
 
     if(options.exportCustomMethods && window.NL_CMETHODS && window.NL_CMETHODS.length > 0) {
-        for(let method of window.NL_CMETHODS) {
+        for(const method of window.NL_CMETHODS) {
             Neutralino.custom[method] = (...args) => {
                 let data = {};
-                for(let [argi, argv] of args.entries()) {
+                for(const [argi, argv] of args.entries()) {
                     if(typeof argv == 'object' && !Array.isArray(argv) && argv != null) {
                         data = {...data, ...argv};
                     }
@@ -43,6 +40,6 @@ export function init(options: InitOptions = {}): void {
     }
 
     window.NL_CVERSION = version;
-    window.NL_CCOMMIT = '<git_commit_hash_latest>'; // only the build server will update this
+    window.NL_CCOMMIT = '425c526c318342e0e5d0f17caceef2a53049eda4'; // only the build server will update this
     initialized = true;
 }