+++ src/api/os.ts
@@ -1,87 +1,16 @@
 import { sendMessage } from '../ws/websocket';
-
-export interface ExecCommandOptions {
-    stdIn?: string;
-    background?: boolean;
-    cwd?: string;
-}
-
-export interface ExecCommandResult {
-    pid: number;
-    stdOut: string;
-    stdErr: string;
-    exitCode: number;
-}
-
-export interface SpawnedProcess {
-    id: number;
-    pid: number;
-}
-
-export interface Envs {
-    [key: string]: string;
-}
-
-export interface OpenDialogOptions {
-    multiSelections?: boolean;
-    filters?: Filter[];
-    defaultPath?: string;
-}
-
-export interface FolderDialogOptions {
-    defaultPath?: string;
-}
-
-export interface SaveDialogOptions {
-    forceOverwrite?: boolean;
-    filters?: Filter[];
-    defaultPath?: string;
-}
-
-export interface Filter {
-    name: string;
-    extensions: string[];
-}
-
-export interface TrayOptions {
-    icon: string;
-    menuItems: TrayMenuItem[];
-}
-
-export interface TrayMenuItem {
-    id?: string;
-    text: string;
-    isDisabled?: boolean;
-    isChecked?: boolean;
-}
-
-export enum Icon {
-    WARNING = 'WARNING',
-    ERROR = 'ERROR',
-    INFO = 'INFO',
-    QUESTION = 'QUESTION'
-};
-
-export enum MessageBoxChoice {
-    OK = 'OK',
-    OK_CANCEL = 'OK_CANCEL',
-    YES_NO = 'YES_NO',
-    YES_NO_CANCEL = 'YES_NO_CANCEL',
-    RETRY_CANCEL = 'RETRY_CANCEL',
-    ABORT_RETRY_IGNORE = 'ABORT_RETRY_IGNORE'
-};
-
-export type KnownPath =
-    'config' |
-    'data' |
-    'cache' |
-    'documents' |
-    'pictures' |
-    'music' |
-    'video' |
-    'downloads' |
-    'savedGames1' |
-    'savedGames2'
+import type { MessageBoxChoice, Icon } from '../types/enums';
+import type {
+    Envs,
+    ExecCommandOptions,
+    ExecCommandResult,
+    FolderDialogOptions,
+    KnownPath,
+    OpenDialogOptions,
+    SaveDialogOptions,
+    SpawnedProcess,
+    TrayOptions
+} from '../types/api/os';
 
 export function execCommand(command: string, options?: ExecCommandOptions): Promise<ExecCommandResult> {
     return sendMessage('os.execCommand', { command, ...options });