+++ src/api/window.ts
@@ -1,38 +1,15 @@
 import { sendMessage } from '../ws/websocket';
 import * as os from './os';
+import {
+    WindowOptions,
+    WindowPosOptions,
+    WindowSizeOptions
+} from '../types/api/window';
 
-const draggableRegions: WeakMap<HTMLElement, any> = new WeakMap();
-
-export interface WindowOptions extends WindowSizeOptions, WindowPosOptions {
-  title?: string;
-  icon?: string;
-  fullScreen?: boolean;
-  alwaysOnTop?: boolean;
-  enableInspector?: boolean;
-  borderless?: boolean;
-  maximize?: boolean;
-  hidden?: boolean;
-  maximizable?: boolean;
-  useSavedState?: boolean;
-  exitProcessOnClose?: boolean;
-  extendUserAgentWith?: string;
-  processArgs?: string;
-}
-
-export interface WindowSizeOptions {
-  width?: number;
-  height?: number;
-  minWidth?: number;
-  minHeight?: number;
-  maxWidth?: number;
-  maxHeight?: number;
-  resizable?: boolean;
-}
-
-export interface WindowPosOptions {
-  x: number;
-  y: number;
-}
+const draggableRegions: WeakMap<HTMLElement, {
+    pointerdown: (e: PointerEvent) => void;
+    pointerup: (e: PointerEvent) => void;
+}> = new WeakMap();
 
 export function setTitle(title: string): Promise<void> {
     return sendMessage('window.setTitle', { title });
@@ -56,6 +33,14 @@ export function isMaximized(): Promise<boolean> {
 
 export function minimize(): Promise<void> {
     return sendMessage('window.minimize');
+};
+
+export function unminimize(): Promise<void> {
+    return sendMessage('window.unminimize');
+};
+
+export function isMinimized(): Promise<boolean> {
+    return sendMessage('window.isMinimized');
 };
 
 export function setFullScreen(): Promise<void> {
@@ -98,15 +83,35 @@ export function center(): Promise<void> {
     return sendMessage('window.center');
 };
 
-export function setDraggableRegion(domElementOrId: string | HTMLElement): Promise<void> {
-    return new Promise((resolve: any, reject: any) => {
+export type DraggableRegionOptions = {
+    /**
+     * If set to `true`, the region will always capture the pointer,
+     * ensuring dragging doesn't break on fast pointer movement.
+     * Note that it prevents child elements from receiving any pointer events.
+     * Defaults to `false`.
+     */
+    alwaysCapture?: boolean;
+    /**
+     * Minimum distance between cursor's starting and current position
+     * after which dragging is started. This helps prevent accidental dragging
+     * while interacting with child elements.
+     * Defaults to `10`. (In pixels.)
+     */
+    dragMinDistance?: number;
+}
+export function setDraggableRegion(domElementOrId: string | HTMLElement, options: DraggableRegionOptions = {}): Promise<{
+    success: true,
+    message: string
+}> {
+    return new Promise<Awaited<ReturnType<typeof setDraggableRegion>>>((resolve, reject) => {
         const draggableRegion: HTMLElement = domElementOrId instanceof Element ?
                                                     domElementOrId : document.getElementById(domElementOrId);
         let initialClientX: number = 0;
         let initialClientY: number = 0;
         let absDragMovementDistance: number = 0;
-        let isPointerCaptured = false;
+        let shouldReposition = false;
         let lastMoveTimestamp = performance.now();
+        let isPointerCaptured = options.alwaysCapture;
 
         if (!draggableRegion) {
             return reject({
@@ -124,12 +129,26 @@ export function setDraggableRegion(domElementOrId: string | HTMLElement): Promis
 
         draggableRegion.addEventListener('pointerdown', startPointerCapturing);
         draggableRegion.addEventListener('pointerup', endPointerCapturing);
+        draggableRegion.addEventListener('pointercancel', endPointerCapturing);
 
         draggableRegions.set(draggableRegion, { pointerdown: startPointerCapturing, pointerup: endPointerCapturing });
 
         async function onPointerMove(evt: PointerEvent) {
+            // Get absolute drag distance from the starting point
+            const dx = evt.clientX - initialClientX,
+                  dy = evt.clientY - initialClientY;
+            absDragMovementDistance = Math.sqrt(dx * dx + dy * dy);
+            // Only start pointer capturing when the user dragged more than a certain amount of distance
+            // This ensures that the user can also click on the dragable area, e.g. if the area is menu / navbar
+            if (absDragMovementDistance >= (options.dragMinDistance ?? 10)) {
+                shouldReposition = true;
+                if (!isPointerCaptured) {
+                    draggableRegion.setPointerCapture(evt.pointerId);
+                    isPointerCaptured = true;
+                }
+            }
 
-            if (isPointerCaptured) {
+            if (shouldReposition) {
 
                 const currentMilliseconds = performance.now();
                 const timeTillLastMove = currentMilliseconds - lastMoveTimestamp;
@@ -149,15 +168,6 @@ export function setDraggableRegion(domElementOrId: string | HTMLElement): Promis
 
                 return;
             }
-
-            // Add absolute drag distance
-            absDragMovementDistance = Math.sqrt(evt.movementX * evt.movementX + evt.movementY * evt.movementY);
-            // Only start pointer capturing when the user dragged more than a certain amount of distance
-            // This ensures that the user can also click on the dragable area, e.g. if the area is menu / navbar
-            if (absDragMovementDistance >= 10) { // TODO: introduce constant instead of magic number?
-                isPointerCaptured = true;
-                draggableRegion.setPointerCapture(evt.pointerId);
-            }
         }
 
         function startPointerCapturing(evt: PointerEvent) {
@@ -165,6 +175,9 @@ export function setDraggableRegion(domElementOrId: string | HTMLElement): Promis
             initialClientX = evt.clientX;
             initialClientY = evt.clientY;
             draggableRegion.addEventListener('pointermove', onPointerMove);
+            if (options.alwaysCapture) {
+                draggableRegion.setPointerCapture(evt.pointerId);
+            }
         }
 
         function endPointerCapturing(evt: PointerEvent) {
@@ -179,8 +192,11 @@ export function setDraggableRegion(domElementOrId: string | HTMLElement): Promis
     });
 };
 
-export function unsetDraggableRegion(domElementOrId: string | HTMLElement): Promise<void> {
-  return new Promise((resolve: any, reject: any) => {
+export function unsetDraggableRegion(domElementOrId: string | HTMLElement): Promise<{
+    success: true,
+    message: string
+}> {
+  return new Promise<Awaited<ReturnType<typeof unsetDraggableRegion>>>((resolve, reject) => {
         const draggableRegion: HTMLElement = domElementOrId instanceof Element ?
                                                 domElementOrId : document.getElementById(domElementOrId);
 
@@ -200,6 +216,7 @@ export function unsetDraggableRegion(domElementOrId: string | HTMLElement): Prom
         const { pointerdown, pointerup } = draggableRegions.get(draggableRegion);
         draggableRegion.removeEventListener('pointerdown', pointerdown);
         draggableRegion.removeEventListener('pointerup', pointerup);
+        draggableRegion.removeEventListener('pointercancel', pointerup);
         draggableRegions.delete(draggableRegion);
 
         resolve({
@@ -267,9 +284,7 @@ export function create(url: string, options?: WindowOptions): Promise<void> {
             if(key == "processArgs")
                 continue;
 
-            let cliKey: string = key.replace(/[A-Z]|^[a-z]/g, (token: string) => (
-               "-" + token.toLowerCase()
-            ));
+            let cliKey: string = '-' + key.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase();
             command += ` --window${cliKey}=${normalize(options[key])}`
         }
 
@@ -285,3 +300,7 @@ export function create(url: string, options?: WindowOptions): Promise<void> {
             });
     });
 };
+
+export function snapshot(path: string): Promise<void> {
+    return sendMessage('window.snapshot', { path });
+}